Reformated TimeStamping in ServiceOne and Two
Messages now include a timestamp of when they are sent. ServiceTwo receiveMessage method then gives over all of that information to another method which processes the received message and timestamp, along with a timestamp of when it was received that is given by receiveMessage. That way, time calculations exclude the time it takes to split the string message and make the actual calculation of serviceOne and Two date formats.